+++ lib/l10n/l10n.dart
@@ -305,7 +305,7 @@ extension L10nProfileTabExtension on ProfileTab {
       ProfileTab.storage => 'label_storage'.l10n,
       ProfileTab.confidential => 'label_confidentiality'.l10n,
       ProfileTab.devices => 'label_devices'.l10n,
-      ProfileTab.download => 'label_download'.l10n,
+      ProfileTab.download => 'label_download_and_update'.l10n,
       ProfileTab.danger => 'label_danger_zone'.l10n,
       ProfileTab.legal => 'label_legal_information'.l10n,
       ProfileTab.support => 'btn_help'.l10n,